Module-18-GenAI-Forge/QA_with_Gmail_Inbox/requirements.txt
gradio
streamlit
PyMuPDF
requests
langchain-openai
python-dotenv
chromadb
langchain
langchain_community
langchain-text-splitters
sentence-transformers
langchain-huggingface
google-api-python-client
google-auth
google-auth-httplib2 
google-auth-oauthlib


